Use Point instead of Coordinate in views
The data model for the views was still using the Coordinate class to
store locations on the plane as a relic from the times when view
locations were geographic locations. Since this has been changed and the
views use a simple cartesian coordinate system now, it makes more sense
to use the Point class to store locations.