Add an abstraction layer for XML DOM parsing
By introducing an abstraction layer for XML DOM access, we will be able
to reuse the XML parsing code on multiple platforms.

The problem is that the classes from the packages javax.xml.parsers and
org.w3c.dom are not available on JSweet and GWT and are not easily
portable. Since we're using only a minimal subset of the XML DOM API
however, it is not difficult to have an abstraction layer for the API
that we actually use and then have different implementations for this layer.

On the desktop, the abstraction layer can be implemented using the
standard parsing facilities from org.w3c.dom and javax.xml.parsers
(done), while on JSweet for example, we will implement it using the JQuery
built-in XML parser.