Add a dot-style station-drawing mode
For maps with many lines running in parallel, the rendering becomes very
cluttered. To overcome the initial mess of imported data with such
outcomes, we need rendering for both stations and lines that displays
such situations clearly. Having a station mode that renders all stations
as disks / dots solves this for the stations.