Move module 'test-gwt' to top level
The Gradle GWT plugin is a bit strange. Running for instance the eclipse
task depends on the compileJava task. This makes it impossible to create
the Eclipse project files in a state where compilation does not work,
which happens often during module refactoring. To avoid this, we move
the GWT testing module up in the directory tree, out of the java modules
directory.